change OtherAssignmentTest
